requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.3.0
pydantic==2.7.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
transformers==4.40.2
